+++ internal/git/remote.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	gogit "github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/transport"
 )
 
@@ -64,4 +65,18 @@ func (r *runner) fetchRemoteShas(repo *Repository, remote string) (map[string]st
 	}
 
 	return remoteShas, nil
+}
+
+func (r *runner) getRemoteSha(repo *Repository, remote, branchName string) (string, error) {
+	// Synchronize go-git operations to prevent concurrent packfile access
+	goGitMu.Lock()
+	defer goGitMu.Unlock()
+
+	refName := plumbing.ReferenceName(fmt.Sprintf("refs/remotes/%s/%s", remote, branchName))
+	ref, err := repo.Reference(refName, true)
+	if err != nil {
+		return "", fmt.Errorf("failed to get remote SHA for %s/%s: %w", remote, branchName, err)
+	}
+
+	return ref.Hash().String(), nil
 }
