+++ internal/git/runner.go
@@ -360,22 +360,6 @@ func (r *runner) EnsureMetadataRefspecConfigured() error {
 	return r.AddConfigValue("remote.origin.fetch", metadataRefspec)
 }
 
-func (r *runner) GetCurrentBranch() (string, error) {
-	repo, err := r.ensureRepo()
-	if err != nil {
-		return "", err
-	}
-	return repo.GetCurrentBranch()
-}
-
-func (r *runner) GetAllBranchNames() ([]string, error) {
-	repo, err := r.ensureRepo()
-	if err != nil {
-		return nil, err
-	}
-	return repo.GetBranchNames()
-}
-
 func (r *runner) FindRemoteBranch(ctx context.Context, remote string) (string, error) {
 	// Get all branch configs that have this remote
 	// Format: "branch.<name>.remote <remote>"
@@ -408,73 +392,12 @@ func (r *runner) FindRemoteBranch(ctx context.Context, remote string) (string, e
 	return "", nil
 }
 
-func (r *runner) CheckoutBranch(ctx context.Context, branchName string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "checkout", branchName)
-	if err != nil {
-		return fmt.Errorf("failed to checkout branch %s: %w", branchName, err)
-	}
-	return nil
-}
-
-func (r *runner) CheckoutBranchForce(ctx context.Context, branchName string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "checkout", "-f", branchName)
-	return err
-}
-
-func (r *runner) CreateAndCheckoutBranch(ctx context.Context, branchName string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "checkout", "-b", branchName)
-	if err != nil {
-		return fmt.Errorf("failed to create and checkout branch %s: %w", branchName, err)
-	}
-	return nil
-}
-
-func (r *runner) DeleteBranch(ctx context.Context, branchName string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "branch", "-D", branchName)
-	if err != nil {
-		return fmt.Errorf("failed to delete branch %s: %w", branchName, err)
-	}
-	return nil
-}
-
-func (r *runner) RenameBranch(ctx context.Context, oldName, newName string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "branch", "-m", oldName, newName)
-	if err != nil {
-		return fmt.Errorf("failed to rename branch %s to %s: %w", oldName, newName, err)
-	}
-	return nil
-}
-
-func (r *runner) CheckoutDetached(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "checkout", "--detach", revision)
-	if err != nil {
-		return fmt.Errorf("failed to checkout %s in detached state: %w", revision, err)
-	}
-	return nil
-}
-
-func (r *runner) UpdateBranchRef(ctx context.Context, branchName, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "update-ref", "refs/heads/"+branchName, revision)
-	if err != nil {
-		return fmt.Errorf("failed to update branch ref: %w", err)
-	}
-	return nil
-}
-
 func (r *runner) GetRemoteRevision(branchName string) (string, error) {
 	repo, err := r.ensureRepo()
 	if err != nil {
 		return "", err
 	}
 	return r.getRemoteRevision(repo, branchName)
-}
-
-func (r *runner) GetCurrentBranchOrSHA(ctx context.Context) (string, error) {
-	branch, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", "--abbrev-ref", "HEAD")
-	if err == nil && branch != "HEAD" {
-		return branch, nil
-	}
-	return r.GetCurrentRevision(ctx)
 }
 
 func (r *runner) GetCurrentRevision(ctx context.Context) (string, error) {
@@ -646,246 +569,6 @@ func (r *runner) GetCommitSHA(branchName string, offset int) (string, error) {
 	return commit.Hash.String(), nil
 }
 
-func (r *runner) PullBranch(ctx context.Context, remote, branchName string) (PullResult, error) {
-	// Save current branch/detached HEAD
-	currentBranch, err := r.GetCurrentBranch()
-	var currentRev string
-	if err != nil {
-		currentBranch = ""
-		currentRev, _ = r.runGitCommandWithContextInternal(ctx, "rev-parse", "HEAD")
-	}
-
-	// Get the SHA of the local branch
-	oldRev, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", branchName)
-	if err != nil {
-		return PullConflict, fmt.Errorf("failed to get local revision for %s: %w", branchName, err)
-	}
-
-	// Fetch first
-	_, _ = r.runGitCommandWithContextInternal(ctx, "fetch", remote, branchName)
-
-	// Get the SHA of the remote branch
-	remoteRev, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", fmt.Sprintf("%s/%s", remote, branchName))
-	if err != nil {
-		// If we can't get remote rev, we can't pull, but it might just be because there's no remote
-		return PullUnneeded, nil //nolint:nilerr
-	}
-
-	if oldRev == remoteRev {
-		return PullUnneeded, nil
-	}
-
-	// Check if it's a fast-forward
-	isAncestor, err := r.IsAncestor(oldRev, remoteRev)
-	if err != nil || !isAncestor {
-		return PullConflict, nil //nolint:nilerr
-	}
-
-	// Update the local branch reference to the remote commit (fast-forward)
-	_, err = r.runGitCommandWithContextInternal(ctx, "update-ref", "refs/heads/"+branchName, remoteRev)
-	if err != nil {
-		return PullConflict, fmt.Errorf("failed to update local branch %s to %s: %w", branchName, remoteRev, err)
-	}
-
-	// If we are currently ON this branch in this worktree, we need to update HEAD
-	if currentBranch == branchName {
-		_ = r.CheckoutBranch(ctx, branchName)
-	} else if currentRev != "" {
-		_ = r.CheckoutDetached(ctx, currentRev)
-	}
-
-	return PullDone, nil
-}
-
-func (r *runner) PushBranch(ctx context.Context, branchName, remote string, opts PushOptions) error {
-	args := []string{"push", "-u", remote}
-
-	if opts.Force {
-		args = append(args, "--force")
-	} else if opts.ForceWithLease {
-		args = append(args, "--force-with-lease")
-	}
-
-	if opts.NoVerify {
-		args = append(args, "--no-verify")
-	}
-
-	args = append(args, branchName)
-
-	_, err := r.runGitCommandWithContextInternal(ctx, args...)
-	if err != nil {
-		if strings.Contains(err.Error(), "stale info") || strings.Contains(err.Error(), "forced update") {
-			return fmt.Errorf("%w: force-with-lease push of %s failed due to external changes to the remote branch", ErrStaleRemoteInfo, branchName)
-		}
-		return fmt.Errorf("failed to push branch %s: %w", branchName, err)
-	}
-
-	return nil
-}
-
-func (r *runner) Rebase(ctx context.Context, branchName, upstream, oldUpstream string) (RebaseResult, error) {
-	// Use detached HEAD to avoid "already used by worktree" errors
-	// We use branchName~0 to force a detached checkout of the branch tip
-	_, err := r.runGitCommandWithContextInternal(ctx, "rebase", "--onto", upstream, oldUpstream, branchName+"~0")
-	if err != nil {
-		if r.IsRebaseInProgress(ctx) {
-			return RebaseConflict, nil
-		}
-		// Abort rebase if it failed for other reasons
-		_, _ = r.runGitCommandWithContextInternal(ctx, "rebase", "--abort")
-
-		return RebaseConflict, err
-	}
-
-	// Since we rebased in detached HEAD, we must manually update the branch ref
-	newRev, err := r.GetCurrentRevision(ctx)
-	if err != nil {
-		return RebaseConflict, fmt.Errorf("failed to get revision after rebase: %w", err)
-	}
-
-	if err := r.UpdateBranchRef(ctx, branchName, newRev); err != nil {
-		return RebaseConflict, fmt.Errorf("failed to update branch ref %s: %w", branchName, err)
-	}
-
-	return RebaseDone, nil
-}
-
-func (r *runner) RebaseContinueNoEdit(ctx context.Context) (RebaseResult, error) {
-	_, err := r.RunGitCommandWithEnv(ctx, []string{"GIT_EDITOR=true"}, "rebase", "--continue")
-	if err != nil {
-		if r.IsRebaseInProgress(ctx) {
-			return RebaseConflict, nil
-		}
-		return RebaseConflict, err
-	}
-	return RebaseDone, nil
-}
-
-func (r *runner) RebaseContinue(ctx context.Context) (RebaseResult, error) {
-	_, err := r.RunGitCommandWithEnv(ctx, []string{"GIT_EDITOR=true"}, "rebase", "--continue")
-	if err != nil {
-		if r.IsRebaseInProgress(ctx) {
-			return RebaseConflict, nil
-		}
-		return RebaseConflict, fmt.Errorf("rebase continue failed: %w", err)
-	}
-
-	return RebaseDone, nil
-}
-
-func (r *runner) RebaseAbort(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "rebase", "--abort")
-	if err != nil {
-		return fmt.Errorf("rebase abort failed: %w", err)
-	}
-	return nil
-}
-
-func (r *runner) InteractiveRebase(_ context.Context, onto string) error {
-	return r.RunGitCommandInteractive("rebase", "-i", onto)
-}
-
-func (r *runner) IsMergeInProgress(_ context.Context) bool {
-	if r.repoRoot == "" {
-		return false
-	}
-	mergeHead := filepath.Join(r.repoRoot, ".git", "MERGE_HEAD")
-	if _, err := os.Stat(mergeHead); err == nil {
-		return true
-	}
-	return false
-}
-
-func (r *runner) MergeAbort(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "merge", "--abort")
-	if err != nil {
-		return fmt.Errorf("merge abort failed: %w", err)
-	}
-	return nil
-}
-
-func (r *runner) CherryPick(ctx context.Context, commitSHA, onto string) (string, error) {
-	if _, err := r.runGitCommandWithContextInternal(ctx, "checkout", "--detach", onto); err != nil {
-		return "", fmt.Errorf("failed to checkout %s: %w", onto, err)
-	}
-
-	if _, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", commitSHA); err != nil {
-		_, _ = r.runGitCommandWithContextInternal(ctx, "cherry-pick", "--abort")
-		return "", fmt.Errorf("failed to cherry-pick %s: %w", commitSHA, err)
-	}
-
-	newSHA, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", "HEAD")
-	if err != nil {
-		return "", fmt.Errorf("failed to get new SHA after cherry-pick: %w", err)
-	}
-
-	return strings.TrimSpace(newSHA), nil
-}
-
-func (r *runner) StashPush(ctx context.Context, message string) (string, error) {
-	args := []string{"stash", "push", "-u"}
-	if message != "" {
-		args = append(args, "-m", message)
-	}
-	output, err := r.runGitCommandWithContextInternal(ctx, args...)
-	if err != nil {
-		return "", fmt.Errorf("stash push failed: %w", err)
-	}
-	return output, nil
-}
-
-func (r *runner) StashPop(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "stash", "pop")
-	if err != nil {
-		return fmt.Errorf("stash pop failed: %w", err)
-	}
-	return nil
-}
-
-func (r *runner) Fetch(ctx context.Context, remote, branch string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "fetch", remote, branch)
-	if err != nil {
-		return fmt.Errorf("failed to fetch %s from %s: %w", branch, remote, err)
-	}
-	return nil
-}
-
-func (r *runner) CreateBranch(ctx context.Context, branchName, startPoint string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "branch", branchName, startPoint)
-	if err != nil {
-		return fmt.Errorf("failed to create branch %s from %s: %w", branchName, startPoint, err)
-	}
-	return nil
-}
-
-func (r *runner) CreateBranchForce(ctx context.Context, branchName, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "branch", "-f", branchName, revision)
-	return err
-}
-
-func (r *runner) Merge(ctx context.Context, branchName string, opts MergeOptions) error {
-	args := []string{"merge"}
-	if opts.FFOnly {
-		args = append(args, "--ff-only")
-	}
-	if opts.NoEdit {
-		args = append(args, "--no-edit")
-	}
-	if opts.NoFF {
-		args = append(args, "--no-ff")
-	}
-	if opts.Message != "" {
-		args = append(args, "-m", opts.Message)
-	}
-	args = append(args, branchName)
-
-	_, err := r.runGitCommandWithContextInternal(ctx, args...)
-	if err != nil {
-		return fmt.Errorf("failed to merge %s: %w", branchName, err)
-	}
-	return nil
-}
-
 func (r *runner) CheckoutPaths(ctx context.Context, branch string, paths []string) error {
 	args := []string{"checkout", branch, "--"}
 	args = append(args, paths...)
@@ -906,349 +589,12 @@ func (r *runner) RemovePaths(ctx context.Context, paths []string) error {
 	return nil
 }
 
-func (r *runner) ResetMerge(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "--merge", revision)
-	if err != nil {
-		return fmt.Errorf("failed to reset --merge to %s: %w", revision, err)
-	}
-	return nil
-}
-
-func (r *runner) HardReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "--hard", revision)
-	if err != nil {
-		return fmt.Errorf("failed to hard reset to %s: %w", revision, err)
-	}
-	return nil
-}
-
-func (r *runner) SoftReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "-q", "--soft", revision)
-	if err != nil {
-		return fmt.Errorf("failed to soft reset to %s: %w", revision, err)
-	}
-	return nil
-}
-
-func (r *runner) MixedReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", revision)
-	return err
-}
-
-func (r *runner) ListStash(ctx context.Context) (string, error) {
-	return r.runGitCommandWithContextInternal(ctx, "stash", "list")
-}
-
 func (r *runner) GetReflog(ctx context.Context, count int, format string) (string, error) {
 	args := []string{"reflog", fmt.Sprintf("-%d", count)}
 	if format != "" {
 		args = append(args, fmt.Sprintf("--format=%s", format))
 	}
 	return r.runGitCommandWithContextInternal(ctx, args...)
-}
-
-func (r *runner) CommitWithOptions(opts CommitOptions) error {
-	args := []string{"commit"}
-
-	if opts.Amend {
-		args = append(args, "--amend")
-	}
-
-	if opts.NoVerify {
-		args = append(args, "--no-verify")
-	}
-
-	if opts.ResetAuthor {
-		args = append(args, "--reset-author")
-	}
-
-	if opts.Verbose > 0 {
-		args = append(args, "-v")
-	}
-
-	if opts.Message != "" {
-		args = append(args, "-m", opts.Message)
-	}
-
-	if opts.NoEdit {
-		args = append(args, "--no-edit")
-	} else if opts.Edit {
-		// Only add -e if explicitly requested (git opens editor by default if no message)
-		args = append(args, "-e")
-	}
-	// If we're in non-interactive mode, or if we have a message and aren't explicitly editing,
-	// use the non-interactive runner.
-	if !utils.IsInteractive() || (opts.Message != "" && !opts.Edit) || (opts.Amend && opts.NoEdit) {
-		_, err := r.runGitCommandWithContextInternal(context.Background(), args...)
-		return err
-	}
-
-	return r.RunGitCommandInteractive(args...)
-}
-
-func (r *runner) Commit(message string, verbose int, noVerify bool) error {
-	return r.CommitWithOptions(CommitOptions{
-		Message:  message,
-		Verbose:  verbose,
-		NoVerify: noVerify,
-	})
-}
-
-func (r *runner) StageAll(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "add", "-A")
-	if err != nil {
-		return fmt.Errorf("failed to stage all changes: %w", err)
-	}
-	return nil
-}
-
-func (r *runner) StagePatch(_ context.Context) error {
-	return r.RunGitCommandInteractive("add", "-p")
-}
-
-func (r *runner) StageTracked(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "add", "-u")
-	if err != nil {
-		return fmt.Errorf("failed to stage tracked changes: %w", err)
-	}
-	return nil
-}
-
-func (r *runner) AddAll(ctx context.Context) error {
-	return r.StageAll(ctx)
-}
-
-func (r *runner) StageChanges(ctx context.Context, opts StagingOptions) error {
-	if opts.Patch && !opts.All {
-		return r.RunGitCommandInteractive("add", "-p")
-	}
-
-	if opts.All {
-		return r.StageAll(ctx)
-	}
-
-	if opts.Update {
-		_, err := r.runGitCommandWithContextInternal(ctx, "add", "-u")
-		return err
-	}
-
-	return nil
-}
-
-func (r *runner) GetRepoInfo(ctx context.Context) (string, string, error) {
-	// Get remote URL
-	url, _ := r.runGitCommandWithContextInternal(ctx, "config", "--get", "remote.origin.url")
-	// url will be empty if there's an error (e.g. remote.origin.url not set)
-	// This happens in many tests and is not a fatal error for most operations.
-	if url == "" {
-		return "", "", nil
-	}
-
-	// Parse URL (handles both https and ssh formats)
-	url = strings.TrimSpace(url)
-	if url == "" {
-		return "", "", nil
-	}
-	url = strings.TrimSuffix(url, ".git")
-	parts := strings.Split(url, "/")
-	if len(parts) < 2 {
-		return "", "", fmt.Errorf("invalid remote URL")
-	}
-
-	repoName := parts[len(parts)-1]
-	var owner string
-	if strings.Contains(url, "@") {
-		// SSH format: git@github.com:owner/repo
-		sshParts := strings.Split(url, ":")
-		if len(sshParts) < 2 {
-			return "", "", fmt.Errorf("invalid SSH remote URL")
-		}
-		pathParts := strings.Split(sshParts[1], "/")
-		if len(pathParts) < 2 {
-			return "", "", fmt.Errorf("invalid SSH remote URL")
-		}
-		owner = pathParts[0]
-	} else {
-		// HTTPS format: https://github.com/owner/repo
-		owner = parts[len(parts)-2]
-	}
-
-	return owner, repoName, nil
-}
-
-func (r *runner) HasStagedChanges(ctx context.Context) (bool, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--cached", "--shortstat")
-	if err != nil {
-		return false, fmt.Errorf("failed to check staged changes: %w", err)
-	}
-	return strings.TrimSpace(output) != "", nil
-}
-
-func (r *runner) HasUnstagedChanges(ctx context.Context) (bool, error) {
-	// Use git diff to check for unstaged changes to tracked files
-	// This is more reliable than parsing porcelain output which gets trimmed
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--name-only")
-	if err != nil {
-		return false, fmt.Errorf("failed to check unstaged changes: %w", err)
-	}
-	return strings.TrimSpace(output) != "", nil
-}
-
-func (r *runner) HasUntrackedFiles(ctx context.Context) (bool, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "ls-files", "--others", "--exclude-standard")
-	if err != nil {
-		return false, fmt.Errorf("failed to check for untracked files: %w", err)
-	}
-	return strings.TrimSpace(output) != "", nil
-}
-
-func (r *runner) IsMerged(ctx context.Context, branchName, target string) (bool, error) {
-	// Get merge base
-	mergeBase, err := r.GetMergeBase(branchName, target)
-	if err != nil {
-		return false, fmt.Errorf("failed to get merge base: %w", err)
-	}
-
-	// Get branch revision
-	branchRev, err := r.GetRevision(branchName)
-	if err != nil {
-		return false, fmt.Errorf("failed to get branch revision: %w", err)
-	}
-
-	// If merge base equals branch revision, branch is already merged
-	if mergeBase == branchRev {
-		return true, nil
-	}
-
-	// Use git cherry to check if all commits are in trunk
-	// git cherry <trunk> <branch> returns commits that are in branch but not in trunk
-	// If empty, all commits are merged
-	cherryOutput, err := r.runGitCommandWithContextInternal(ctx, "cherry", target, branchName)
-	if err != nil {
-		// If cherry fails, fall back to simpler check
-		// Check if branch tip is reachable from trunk
-		return r.IsAncestor(branchRev, target)
-	}
-
-	// If cherry output is empty or all lines start with '-', branch is merged
-	if cherryOutput == "" {
-		return true, nil
-	}
-
-	// Check if all commits are marked as merged (lines starting with '-')
-	lines := strings.Split(strings.TrimSpace(cherryOutput), "\n")
-	for _, line := range lines {
-		if line != "" && line[0] != '-' {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func (r *runner) GetMergedBranches(ctx context.Context, target string) (map[string]bool, error) {
-	out, err := r.runGitCommandWithContextInternal(ctx, "branch", "--merged", target)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get merged branches: %w", err)
-	}
-
-	merged := make(map[string]bool)
-	lines := strings.Split(out, "\n")
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line == "" {
-			continue
-		}
-		// Remove current branch indicator '*' if present
-		line = strings.TrimPrefix(line, "*")
-		line = strings.TrimSpace(line)
-		merged[line] = true
-	}
-	return merged, nil
-}
-
-func (r *runner) IsDiffEmpty(ctx context.Context, branchName, base string) (bool, error) {
-	branchRev, err := r.GetRevision(branchName)
-	if err != nil {
-		return false, fmt.Errorf("failed to get branch revision: %w", err)
-	}
-
-	if branchRev == base {
-		return true, nil
-	}
-
-	_, err = r.runGitCommandWithContextInternal(ctx, "diff", "--quiet", base, branchRev)
-	return err == nil, nil
-}
-
-func (r *runner) GetChangedFiles(ctx context.Context, base, head string) ([]string, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--name-only", base, head)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get changed files: %w", err)
-	}
-	if output == "" {
-		return []string{}, nil
-	}
-	return strings.Split(strings.TrimSpace(output), "\n"), nil
-}
-
-func (r *runner) ShowDiff(ctx context.Context, left, right string, stat bool) (string, error) {
-	args := []string{"-c", "color.ui=always", "--no-pager", "diff", "--no-ext-diff"}
-	if stat {
-		args = append(args, "--stat")
-	}
-	args = append(args, left, right, "--")
-	return r.runGitCommandWithContextInternal(ctx, args...)
-}
-
-func (r *runner) ShowCommits(ctx context.Context, base, head string, patch, stat bool) (string, error) {
-	args := []string{"-c", "color.ui=always", "--no-pager", "log"}
-	switch {
-	case patch && stat:
-		args = append(args, "--stat")
-	case patch:
-		args = append(args, "-p")
-	default:
-		args = append(args, "--pretty=format:%h - %s")
-	}
-
-	// If base is empty, use head~ (parent commit) for trunk
-	baseRef := base
-	if base == "" {
-		baseRef = head + "~"
-	}
-	args = append(args, fmt.Sprintf("%s..%s", baseRef, head))
-	args = append(args, "--")
-	return r.runGitCommandWithContextInternal(ctx, args...)
-}
-
-func (r *runner) GetUnmergedFiles(ctx context.Context) ([]string, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--name-only", "--diff-filter=U")
-	if err != nil {
-		return []string{}, nil //nolint:nilerr
-	}
-	if output == "" {
-		return []string{}, nil
-	}
-	return strings.Split(strings.TrimSpace(output), "\n"), nil
-}
-
-func (r *runner) GetStagedDiff(ctx context.Context, files ...string) (string, error) {
-	args := []string{"diff", "--cached"}
-	if len(files) > 0 {
-		args = append(args, "--")
-		args = append(args, files...)
-	}
-	return r.runGitCommandRawWithContextInternal(ctx, args...)
-}
-
-func (r *runner) GetUnstagedDiff(ctx context.Context, files ...string) (string, error) {
-	args := []string{"diff"}
-	if len(files) > 0 {
-		args = append(args, "--")
-		args = append(args, files...)
-	}
-	return r.runGitCommandRawWithContextInternal(ctx, args...)
 }
 
 func (r *runner) GetDiffNumstat(base, head string) (string, error) {
@@ -1282,60 +628,6 @@ func (r *runner) GetCommitTemplate(ctx context.Context) (string, error) {
 	}
 
 	return sb.String(), nil
-}
-
-func (r *runner) ParseStagedHunks(ctx context.Context) ([]Hunk, error) {
-	diffOutput, err := r.runGitCommandRawWithContextInternal(ctx, "diff", "--cached")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get staged diff: %w", err)
-	}
-
-	return ParseDiffOutput(diffOutput)
-}
-
-func (r *runner) AddWorktree(ctx context.Context, path string, branch string, detach bool) error {
-	args := []string{"worktree", "add"}
-	if detach {
-		args = append(args, "--detach")
-	}
-	args = append(args, path)
-	if branch != "" {
-		args = append(args, branch)
-	}
-
-	_, err := r.runGitCommandWithContextInternal(ctx, args...)
-	if err != nil {
-		return fmt.Errorf("failed to add worktree at %s: %w", path, err)
-	}
-	return nil
-}
-
-func (r *runner) RemoveWorktree(ctx context.Context, path string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "worktree", "remove", "--force", path)
-	if err != nil {
-		return fmt.Errorf("failed to remove worktree at %s: %w", path, err)
-	}
-	return nil
-}
-
-func (r *runner) ListWorktrees(ctx context.Context) ([]string, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "worktree", "list", "--porcelain")
-	if err != nil {
-		return nil, fmt.Errorf("failed to list worktrees: %w", err)
-	}
-
-	if output == "" {
-		return []string{}, nil
-	}
-
-	lines := strings.Split(strings.TrimSpace(output), "\n")
-	var worktrees []string
-	for _, line := range lines {
-		if len(line) > 9 && line[:9] == "worktree " {
-			worktrees = append(worktrees, line[9:])
-		}
-	}
-	return worktrees, nil
 }
 
 func (r *runner) runGitCommandInternal(args ...string) (string, error) {
@@ -1518,60 +810,6 @@ func (r *runner) CheckCommutation(hunk Hunk, commitSHA, parentSHA string) (bool,
 	return true, nil
 }
 
-func (r *runner) GetRebaseHead() (string, error) {
-	// Try standard rebase head refs in order:
-	// 1. REBASE_HEAD (standard)
-	// 2. refs/rebase-merge/head (interactive)
-	// 3. refs/rebase-apply/head (non-interactive)
-	refs := []string{
-		"REBASE_HEAD",
-		"refs/rebase-merge/head",
-		"refs/rebase-apply/head",
-	}
-
-	for _, refName := range refs {
-		output, err := r.runGitCommandInternal("rev-parse", "--verify", refName)
-		if err == nil && output != "" {
-			return strings.TrimSpace(output), nil
-		}
-	}
-
-	return "", fmt.Errorf("rebase head not found")
-}
-
-func (r *runner) IsRebaseInProgress(ctx context.Context) bool {
-	output, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", "--git-dir")
-	if err != nil {
-		return false
-	}
-
-	gitDir := strings.TrimSpace(output)
-	if _, err := os.Stat(gitDir + "/rebase-merge"); err == nil {
-		return true
-	}
-	if _, err := os.Stat(gitDir + "/rebase-apply"); err == nil {
-		return true
-	}
-	return false
-}
-
-func (r *runner) CheckRebaseInProgress(ctx context.Context) error {
-	if r.IsRebaseInProgress(ctx) {
-		return fmt.Errorf("a rebase is already in progress. Please finish or abort it first")
-	}
-	return nil
-}
-
-func (r *runner) CherryPickSimple(ctx context.Context, commitSHA string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", commitSHA)
-	return err
-}
-
-func (r *runner) CherryPickAbort(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", "--abort")
-	return err
-}
-
 func (r *runner) ApplyPatch(ctx context.Context, patchFile string, threeWay bool) error {
 	args := []string{"apply"}
 	if threeWay {
@@ -1579,11 +817,6 @@ func (r *runner) ApplyPatch(ctx context.Context, patchFile string, threeWay bool
 	}
 	args = append(args, patchFile)
 	_, err := r.runGitCommandWithContextInternal(ctx, args...)
-	return err
-}
-
-func (r *runner) CommitAmendNoEdit(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "commit", "-a", "--amend", "--no-edit", "--no-verify")
 	return err
 }
